Full Information Maximum Likelihood Linear Regression Whether Covariates Include Missing Values or Response Include Missing Values.

Estimate regression coefficients based on Full Information Maximum Likelihood Estimation, which can couple missing data, including response missing or covariates missing.

# S3 method for 'formula'
  fimlreg(formula, data, ...)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under 'Details'.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which clse is called.

...
Optional arguments to be passed to the stats::nlm optimization routine.

Details

Note that arguments ... of stats::nlm are the parameters of algorithm, see the details in help file of "nlm". "fimlreg" can cople with any type of missing data.

Value

Return a list including following components:

beta

A named vector of coefficients

iterations

A positive integer, iterative times in total.

stop.code

The stop code returned by nlm.

Author

Liu Wei

See also

Examples

data(nhanes)
## example one: include missing value
fimlreg(age~., data=nhanes, iterlim= 40)

#> $beta
#>   intercept         bmi         hyp         chl 
#>  2.04957444 -0.11971495  0.44622250  0.01204298 
#> 
#> $iterations
#> [1] 40
#> 
#> $stop.code
#> [1] 4
#> 
#> $formula
#> age ~ .
#> <environment: 0x000000002c2c34a8>
#> 
#> $data
#>    age  bmi hyp chl
#> 1    1   NA  NA  NA
#> 2    2 22.7   1 187
#> 3    1   NA   1 187
#> 4    3   NA  NA  NA
#> 5    1 20.4   1 113
#> 6    3   NA  NA 184
#> 7    1 22.5   1 118
#> 8    1 30.1   1 187
#> 9    2 22.0   1 238
#> 10   2   NA  NA  NA
#> 11   1   NA  NA  NA
#> 12   2   NA  NA  NA
#> 13   3 21.7   1 206
#> 14   2 28.7   2 204
#> 15   1 29.6   1  NA
#> 16   1   NA  NA  NA
#> 17   3 27.2   2 284
#> 18   2 26.3   2 199
#> 19   1 35.3   1 218
#> 20   3 25.5   2  NA
#> 21   1   NA  NA  NA
#> 22   1 33.2   1 229
#> 23   1 27.5   1 131
#> 24   3 24.9   1  NA
#> 25   2 27.4   1 186
#> 
#> attr(,"class")
#> [1] "fiml"

# example two: No missing vlaue
## example two: No missing value
n <- 100
group <- rnorm(n, sd=4)
weight <- 3.2*group + 1.5 + rnorm(n, sd=0.1)
fimllm <- fimlreg(weight~group, data=data.frame(weight=weight, group=group))
fimllm

#> $beta
#> intercept     group 
#>  1.504971  3.198001 
#> 
#> $iterations
#> [1] 1
#> 
#> $stop.code
#> [1] 2
#> 
#> $formula
#> weight ~ group
#> <environment: 0x000000002c2c34a8>
#> 
#> $data
#>            weight       group
#> 1     1.787161061  0.09103879
#> 2    -1.619747046 -0.99771150
#> 3   -26.276344437 -8.68927648
#> 4   -15.288942169 -5.35863647
#> 5   -10.974643351 -3.92194302
#> 6   -21.407579924 -7.17112270
#> 7     8.274938532  2.09125124
#> 8   -18.899518152 -6.36531391
#> 9    11.891178431  3.22505253
#> 10   -7.208945893 -2.73525227
#> 11    5.815989503  1.33755223
#> 12   13.558058427  3.77202569
#> 13  -10.068683834 -3.59355070
#> 14   11.379333761  3.06934347
#> 15    2.660219105  0.36935471
#> 16   -2.697951014 -1.40678410
#> 17    2.817607749  0.46149202
#> 18   -0.401114237 -0.57304541
#> 19  -29.400843143 -9.67928138
#> 20  -13.511719877 -4.65517614
#> 21    2.708783635  0.41370661
#> 22   -0.127842104 -0.50457331
#> 23   -0.372420286 -0.56697364
#> 24   13.319729054  3.64174814
#> 25   20.530304195  5.91047166
#> 26   22.648893871  6.58497041
#> 27   -0.820222297 -0.71997210
#> 28   14.131288265  3.90742554
#> 29   -0.870426203 -0.74641094
#> 30    1.210976620 -0.13639357
#> 31   20.404028726  5.95492881
#> 32   -0.899502985 -0.75935719
#> 33  -12.432416355 -4.35723847
#> 34    6.127466118  1.48435793
#> 35  -11.610605455 -4.11108045
#> 36    1.741729954  0.05966339
#> 37   -5.124800330 -2.08119425
#> 38    0.907092692 -0.20897771
#> 39  -16.839925642 -5.74031829
#> 40   -2.695846077 -1.31191626
#> 41  -14.973263838 -5.14139048
#> 42   -3.973237820 -1.70214896
#> 43    4.597274882  1.02990264
#> 44    4.041126245  0.80527758
#> 45    7.625161003  1.92830961
#> 46    8.013564121  2.02654636
#> 47   -5.270567312 -2.10676543
#> 48   -5.677916484 -2.25909592
#> 49    1.906661399  0.16159214
#> 50   -1.092249140 -0.80312552
#> 51  -10.219195237 -3.61686878
#> 52    3.673120584  0.66662210
#> 53   -5.067988109 -2.09713913
#> 54    1.676603859  0.07025561
#> 55   13.740258815  3.79406875
#> 56    4.843031976  1.08106314
#> 57   -0.397971059 -0.64376562
#> 58   26.643808877  7.86079803
#> 59    6.485694267  1.58219217
#> 60   12.609980060  3.49487748
#> 61    7.627319860  1.88929662
#> 62  -17.290594906 -5.94090580
#> 63   -9.269982089 -3.34464920
#> 64   -6.814404873 -2.58507685
#> 65    0.815031019 -0.18974815
#> 66    3.703482091  0.67103995
#> 67    0.001094333 -0.39895468
#> 68    2.531213262  0.34398582
#> 69   -3.307302596 -1.49029613
#> 70    4.026913248  0.72533017
#> 71  -10.217920885 -3.62722564
#> 72    8.971217674  2.36462861
#> 73   11.580592535  3.17005293
#> 74    2.560642254  0.30748953
#> 75    0.700379659 -0.27134880
#> 76    7.218229677  1.73315219
#> 77    2.213552084  0.19638161
#> 78    1.123150215 -0.13120019
#> 79   -5.145364906 -2.04369910
#> 80    6.057722796  1.42572215
#> 81    6.836522400  1.67178454
#> 82    8.881880980  2.31682104
#> 83  -17.334670157 -5.90063462
#> 84   18.525958561  5.29522092
#> 85   14.717615994  4.12248586
#> 86    5.347509242  1.26949557
#> 87  -12.663357636 -4.44761515
#> 88    9.274258583  2.48484468
#> 89   24.711218867  7.23643421
#> 90   15.760915459  4.45594402
#> 91    7.381616616  1.86140214
#> 92  -12.257892630 -4.27451891
#> 93    4.806114071  1.02376113
#> 94   -8.442935767 -3.10887574
#> 95  -10.627650525 -3.80127295
#> 96   17.134090379  4.92206533
#> 97   -2.190057442 -1.16128532
#> 98  -14.680906320 -4.98099834
#> 99  -10.226507469 -3.70044249
#> 100  -2.881847946 -1.40703308
#> 
#> attr(,"class")
#> [1] "fiml"