Linear Regression by Iterative Least Square Estimation

Linear regression when covariates include missing values embedding the correlation imformation bettwen covariates by Iterative Least Square Estimation.

ilse(formula, data,  bw=NULL, intercept=F,k.type=NULL,bw.type='fix.bw',K=NULL,method="Par.cond", max.iter=50, peps=1e-7, feps = 1e-7,infor_output=F)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under 'Details'.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which ilse is called.

bw
a positive value, specify the bandwidth in estimating missing values, default as NULL. When bw=NULL, it is automatically selected by empirical method.

intercept
a optional logical value, indicate whether X or data including intercept term (full-one column), default as 'FALSE'.

k.type
an optional character string, specify the type of kernel used in iterative estimating algorithm and support 'epk', 'biweight', 'triangle', 'guassian', 'triweight', 'tricube', 'cosine', 'uniform' in current version, defualt as 'guassian'.

bw.type
an optional character string, specify the type of kernel bandwidth, only support 'fix.bw'(fixed bandwidth) in current version, may spport variable bandwidth in future version, defualt as 'fix.bw'.

K
an optional positive integer, specify the sample size used in kernel estimates when choose the variable bandwidth.

method
an optional character string, specify the iterative algorithm, only support 'Par.cond' and 'Full.cond' in current version and will support 'SGD' in future version.

max.iter
an optional positive integer, the maximum iterative times, defualt as '50'.

peps
an optional positive value, tolerance vlaue of relative variation rate of estimated parametric vector, defualt as '1e-7'.

feps
an optional positive vlaue, tolerance vlaue of relative variation rate of objective function value, defualt as '1e-7'.

infor_output
an optional logical value, indicate whether output the iterative information, default as 'TRUE'.

Details

Models for ilse are specified symbolically. A typical model has the form response ~ terms where response is the (numeric) response vector and terms is a series of terms which specifies a linear predictor for response. A terms specification of the form first + second indicates all the terms in first together with all the terms in second with duplicates removed. A specification of the form first:second indicates the set of terms obtained by taking the interactions of all terms in first with all terms in second. The specification first*second indicates the cross of first and second. This is the same as first + second + first:second.

Value

ilse returns an object of class "NAlm".

The functions summary and anova are used to obtain and print a summary and analysis of variance table of the results. The generic accessor functions coefficients, effects, fitted.values and residuals extract various useful features of the value returned by lm.

An object of class "NAlm" is a list containing at least the following components:

beta

a named vector of coefficients

Bmat

a named matrix that summary the estimated beta in every iteration.

d.fn

a nonnegative value, vlaue of relative variation rate of objective function value

d.par

a nonnegative value, relative variation rate of estimated parametric vector when algorithm stopped.

iterations

a positive integer, iterative times in total.

residuals

the residuals, that is response minus fitted values.

fitted.values

the fitted mean values.

inargs

a list including all input arguments.

References

Huazhen Lin, Wei Liu, Wei Lan(2019). Regression Analysis with individual-specific patterns of missing covariates.

Author

Wei Liu

Note

nothing

See also

Examples

## exmaple one: include missing value
data(nhanes)
NAlm1 <- ilse(age~., data=nhanes, method = 'Full.cond')
NAlm1

#> $beta
#>   intercept         bmi         hyp         chl 
#>  2.04634007 -0.11517617  0.51620712  0.01118342 
#> 
#> $Bmat
#>           [,1]       [,2]      [,3]       [,4]
#> iter0 1.773958 -0.1040127 0.4108214 0.01129992
#> iter1 2.050373 -0.1152149 0.5131551 0.01118582
#> iter2 2.046408 -0.1151781 0.5161661 0.01118350
#> iter3 2.046340 -0.1151762 0.5162068 0.01118342
#> iter4 2.046340 -0.1151762 0.5162071 0.01118342
#> 
#> $d.fn
#> [1] 1.271989e-07
#> 
#> $d.par
#> [1] 1.725446e-07
#> 
#> $iterations
#> [1] 4
#> 
#> $residuals
#>           1           2           3           4           5           6 
#> -0.76520991 -0.01847994 -0.54533187  1.23479009 -0.50728378  1.37869100 
#>           7           8           9          10          11          12 
#> -0.31785603 -0.16618742 -0.63398047  0.23479009 -0.76520991  0.23479009 
#>          13          14          15          16          17          18 
#>  0.66707702 -0.12386915 -0.26632962 -0.76520991 -0.13565588 -0.34784930 
#>          19          20          21          22          23          24 
#>  0.10759875  0.65511924 -0.76520991 -0.24963395  0.12167577  1.20543619 
#>          25 
#>  0.53332879 
#> 
#> $fitted.values
#>         1         2         3         4         5         6         7         8 
#> 1.7652099 2.0184799 1.5453319 1.7652099 1.5072838 1.6213090 1.3178560 1.1661874 
#>         9        10        11        12        13        14        15        16 
#> 2.6339805 1.7652099 1.7652099 1.7652099 2.3329230 2.1238691 1.2663296 1.7652099 
#>        17        18        19        20        21        22        23        24 
#> 3.1356559 2.3478493 0.8924013 2.3448808 1.7652099 1.2496339 0.8783242 1.7945638 
#>        25 
#> 1.4666712 
#> 
#> $inargs
#> $inargs$bw
#> [1] 1
#> 
#> $inargs$intercept
#> [1] FALSE
#> 
#> $inargs$k.type
#> NULL
#> 
#> $inargs$bw.type
#> [1] "fix.bw"
#> 
#> $inargs$K
#> NULL
#> 
#> $inargs$method
#> [1] "Full.cond"
#> 
#> $inargs$max.iter
#> [1] 50
#> 
#> $inargs$peps
#> [1] 1e-05
#> 
#> $inargs$feps
#> [1] 1e-07
#> 
#> $inargs$infor_output
#> [1] FALSE
#> 
#> 
#> $call
#> clse(formula = age ~ ., data = nhanes, method = "Full.cond")
#> 
#> $formula
#> age ~ .
#> <environment: 0x000000002c0f8218>
#> 
#> $data
#>    age  bmi hyp chl
#> 1    1   NA  NA  NA
#> 2    2 22.7   1 187
#> 3    1   NA   1 187
#> 4    3   NA  NA  NA
#> 5    1 20.4   1 113
#> 6    3   NA  NA 184
#> 7    1 22.5   1 118
#> 8    1 30.1   1 187
#> 9    2 22.0   1 238
#> 10   2   NA  NA  NA
#> 11   1   NA  NA  NA
#> 12   2   NA  NA  NA
#> 13   3 21.7   1 206
#> 14   2 28.7   2 204
#> 15   1 29.6   1  NA
#> 16   1   NA  NA  NA
#> 17   3 27.2   2 284
#> 18   2 26.3   2 199
#> 19   1 35.3   1 218
#> 20   3 25.5   2  NA
#> 21   1   NA  NA  NA
#> 22   1 33.2   1 229
#> 23   1 27.5   1 131
#> 24   3 24.9   1  NA
#> 25   2 27.4   1 186
#> 
#> attr(,"class")
#> [1] "ilse"

NAlm2 <- ilse(age~., data=nhanes,bw=1,bw.type = 'fix.bw',method = 'Par.cond', k.type='gaussian', infor_output = T)

#> iter:0d.fn:NAd.par:NA

#> par:1.77-0.10.410.01

#> iter:1d.fn:0.00126549287943754d.par:0.158800467611634

#> coefs:2.06-0.120.620.01

#> iter:2d.fn:0.00199253706347666d.par:0.00219612310896849

#> coefs:2.05-0.110.620.01

#> iter:3d.fn:2.59094226751189e-05d.par:2.91764380510912e-05

#> coefs:2.05-0.110.620.01

#> iter:4d.fn:3.85951952865251e-07d.par:3.20422782331222e-07

#> coefs:2.05-0.110.620.01

## example two: No missing value
n <- 100
group <- rnorm(n, sd=4)
weight <- 3.2*group + 1.5 + rnorm(n, sd=0.1)
NAlm3 <- ilse(weight~group, data=data.frame(weight=weight, group=group), intercept = F)
NAlm3

#> $beta
#> intercept     group 
#>  1.492943  3.198031 
#> 
#> $Bmat
#>           [,1]     [,2]
#> iter0 1.492943 3.198031
#> iter1 1.492943 3.198031
#> 
#> $d.fn
#> [1] 0
#> 
#> $d.par
#> [1] 0
#> 
#> $iterations
#> [1] 1
#> 
#> $residuals
#>            1            2            3            4            5            6 
#>  0.021348644 -0.016931183 -0.135621509  0.124327475  0.025717648  0.045299851 
#>            7            8            9           10           11           12 
#>  0.042226617  0.067103317  0.047893000  0.135955929  0.067310672 -0.018016112 
#>           13           14           15           16           17           18 
#>  0.037923428 -0.249757447  0.114894137  0.020889714 -0.134906394  0.046984303 
#>           19           20           21           22           23           24 
#>  0.105801750  0.036893870 -0.008975084 -0.029406094  0.154011720 -0.083842579 
#>           25           26           27           28           29           30 
#>  0.031230245  0.035142202 -0.209946101  0.112695029  0.062659916  0.029659305 
#>           31           32           33           34           35           36 
#> -0.181492425  0.016007123 -0.096221963  0.192066167  0.142228363 -0.075127282 
#>           37           38           39           40           41           42 
#>  0.061830301  0.090571403 -0.030365118  0.030136798 -0.098025713 -0.127419226 
#>           43           44           45           46           47           48 
#>  0.105237917 -0.063219186 -0.081396978  0.139261594 -0.121965396  0.052666561 
#>           49           50           51           52           53           54 
#> -0.064995018 -0.256957068  0.286883573  0.043723643 -0.048703645  0.006047338 
#>           55           56           57           58           59           60 
#> -0.002938833 -0.007950348 -0.008513384 -0.164739876 -0.087994266 -0.040051325 
#>           61           62           63           64           65           66 
#>  0.030375641 -0.015718579 -0.022239392  0.046338195  0.029218312 -0.221248988 
#>           67           68           69           70           71           72 
#>  0.007246629 -0.019856219  0.035411374 -0.178033957 -0.027497404  0.123836645 
#>           73           74           75           76           77           78 
#> -0.057810350  0.078906987 -0.056590580 -0.036755323  0.059554357  0.064816429 
#>           79           80           81           82           83           84 
#>  0.072459325 -0.085565245  0.002931726 -0.139982992  0.033252543  0.125502810 
#>           85           86           87           88           89           90 
#> -0.047652145 -0.008215222  0.105736619 -0.030920015  0.207499999 -0.046976782 
#>           91           92           93           94           95           96 
#> -0.136206567  0.181826016  0.076484440 -0.076706485 -0.077447902 -0.065165605 
#>           97           98           99          100 
#> -0.039984179 -0.011202839  0.012235847  0.020992873 
#> 
#> $fitted.values
#>           1           2           3           4           5           6 
#>  -9.7973302  -4.7753323  23.9192353 -10.4076070   6.7104721   8.2688555 
#>           7           8           9          10          11          12 
#>  14.8426841  -1.3127210 -18.1961388  -3.9574415  -0.9931108  -2.6712333 
#>          13          14          15          16          17          18 
#>  10.8312868  -6.7093808  10.7905743   1.9700333   9.4432774  17.6742153 
#>          19          20          21          22          23          24 
#>  33.9258626 -16.6234956   9.4203531 -10.7199952   8.8246580 -20.8373030 
#>          25          26          27          28          29          30 
#> -12.3751460 -13.9800451   8.3310731 -15.3294672   0.3913476  -6.3923193 
#>          31          32          33          34          35          36 
#>  25.6501401  -3.1358843 -18.4916778  22.3129856  -3.4414407  -8.6539800 
#>          37          38          39          40          41          42 
#>  -0.7096777  -6.6468591  -7.7130562 -23.5472156 -11.1641074   5.5695714 
#>          43          44          45          46          47          48 
#>  15.2936608 -11.1416911 -15.5295778   8.3828295  24.1531686  11.0847163 
#>          49          50          51          52          53          54 
#>   1.0276458  12.8567905   1.8864106  -0.5763009 -17.9148860  10.0491589 
#>          55          56          57          58          59          60 
#>  -9.0733872  12.3446965  13.4222878  -7.7754536 -26.8328382  -6.6967375 
#>          61          62          63          64          65          66 
#>   6.8790688 -10.0155873   7.0161049  -7.5574609  22.4637086  22.0667039 
#>          67          68          69          70          71          72 
#>  15.8152517  -6.8981200  -7.8540059 -10.5182514  -7.0850845  -2.6708423 
#>          73          74          75          76          77          78 
#>  -8.7524848 -14.9376583  18.5003587  -7.4017131  12.6533691   0.6067947 
#>          79          80          81          82          83          84 
#>  14.3315885   0.7787594  22.7285509   9.2924408  14.4945914 -17.9871248 
#>          85          86          87          88          89          90 
#>  15.9682273   3.7987013  12.9234990   3.6446692   0.7101761   5.9220286 
#>          91          92          93          94          95          96 
#>  -8.2223783 -11.7058999  20.8087064  -0.2800254  -0.2560019  12.7545584 
#>          97          98          99         100 
#> -15.6596750  -1.8780949  -4.6667332  -2.1332247 
#> 
#> $inargs
#> $inargs$bw
#> [1] 2.745281
#> 
#> $inargs$intercept
#> [1] FALSE
#> 
#> $inargs$k.type
#> NULL
#> 
#> $inargs$bw.type
#> [1] "fix.bw"
#> 
#> $inargs$K
#> NULL
#> 
#> $inargs$method
#> [1] "Par.cond"
#> 
#> $inargs$max.iter
#> [1] 50
#> 
#> $inargs$peps
#> [1] 1e-05
#> 
#> $inargs$feps
#> [1] 1e-07
#> 
#> $inargs$infor_output
#> [1] FALSE
#> 
#> 
#> $call
#> clse(formula = weight ~ group, data = data.frame(weight = weight, 
#>     group = group), intercept = F)
#> 
#> $formula
#> weight ~ group
#> <environment: 0x000000002c0f8218>
#> 
#> $data
#>          weight      group
#> 1    -9.7759816 -3.5303824
#> 2    -4.7922635 -1.9600419
#> 3    23.7836138  7.0125307
#> 4   -10.2832795 -3.7212113
#> 5     6.7361898  1.6314816
#> 6     8.3141553  2.1187762
#> 7    14.8849108  4.1743623
#> 8    -1.2456176 -0.8773098
#> 9   -18.1482458 -6.1566259
#> 10   -3.8214855 -1.7042937
#> 11   -0.9258001 -0.7773701
#> 12   -2.6892495 -1.3021062
#> 13   10.8692102  2.9200289
#> 14   -6.9591383 -2.5648042
#> 15   10.9054684  2.9072984
#> 16    1.9909231  0.1491825
#> 17    9.3083710  2.4860090
#> 18   17.7211996  5.0597605
#> 19   34.0316644 10.1415268
#> 20  -16.5866017 -5.6648723
#> 21    9.4113781  2.4788407
#> 22  -10.7494013 -3.8188928
#> 23    8.9786697  2.2925714
#> 24  -20.9211456 -6.9824978
#> 25  -12.3439157 -4.3364458
#> 26  -13.9449029 -4.8382855
#> 27    8.1211270  2.1382311
#> 28  -15.2167722 -5.2602395
#> 29    0.4540075 -0.3444605
#> 30   -6.3626600 -2.4656615
#> 31   25.4686477  7.5537715
#> 32   -3.1198772 -1.4473991
#> 33  -18.5878998 -6.2490387
#> 34   22.5050518  6.5102687
#> 35   -3.2992124 -1.5429442
#> 36   -8.7291073 -3.1728655
#> 37   -0.6478474 -0.6887427
#> 38   -6.5562877 -2.5452541
#> 39   -7.7434213 -2.8786458
#> 40  -23.5170788 -7.8298668
#> 41  -11.2621331 -3.9577633
#> 42    5.4421522  1.2747306
#> 43   15.3988987  4.3153793
#> 44  -11.2049103 -3.9507538
#> 45  -15.6109748 -5.3228126
#> 46    8.5220911  2.1544150
#> 47   24.0312032  7.0856799
#> 48   11.1373828  2.9992744
#> 49    0.9626508 -0.1454949
#> 50   12.5998334  3.5533885
#> 51    2.1732941  0.1230343
#> 52   -0.5325772 -0.6470368
#> 53  -17.9635897 -6.0686803
#> 54   10.0552062  2.6754635
#> 55   -9.0763260 -3.3040110
#> 56   12.3367461  3.3932606
#> 57   13.4137744  3.7302152
#> 58   -7.9401935 -2.8981570
#> 59  -26.9208325 -8.8572560
#> 60   -6.7367888 -2.5608507
#> 61    6.9094444  1.6842005
#> 62  -10.0313059 -3.5986298
#> 63    6.9938655  1.7270506
#> 64   -7.5111227 -2.8299924
#> 65   22.4929270  6.5573986
#> 66   21.8454549  6.4332582
#> 67   15.8224983  4.4784768
#> 68   -6.9179762 -2.6238215
#> 69   -7.8185945 -2.9227197
#> 70  -10.6962854 -3.7558090
#> 71   -7.1125819 -2.6822839
#> 72   -2.5470057 -1.3019840
#> 73   -8.8102952 -3.2036672
#> 74  -14.8587514 -5.1377239
#> 75   18.4437681  5.3180893
#> 76   -7.4384684 -2.7812912
#> 77   12.7129235  3.4897802
#> 78    0.6716111 -0.2770919
#> 79   14.4040479  4.0145466
#> 80    0.6931942 -0.2233198
#> 81   22.7314827  6.6402128
#> 82    9.1524578  2.4388435
#> 83   14.5278439  4.0655164
#> 84  -17.8616220 -6.0912688
#> 85   15.9205751  4.5263111
#> 86    3.7904861  0.7209930
#> 87   13.0292356  3.5742477
#> 88    3.6137491  0.6728284
#> 89    0.9176761 -0.2447653
#> 90    5.8750519  1.3849413
#> 91   -8.3585849 -3.0379070
#> 92  -11.5240739 -4.1271776
#> 93   20.8851908  6.0398920
#> 94   -0.3567319 -0.5543937
#> 95   -0.3334498 -0.5468818
#> 96   12.6893928  3.5214213
#> 97  -15.6996592 -5.3634930
#> 98   -1.8892977 -1.0540979
#> 99   -4.6544973 -1.9260838
#> 100  -2.1122318 -1.1338750
#> 
#> attr(,"class")
#> [1] "ilse"