print.fiml {ILSE}	R Documentation

Print the Information of FIML Regression Whether Covariates Include Missing Values or Response Include Missing Values.

Description

print method for class "fiml"..

Usage


  ## S3 method for class 'fiml'
  print(object)

Arguments

object
an object of class "fiml".

Value

print the basic information of fiml estimation and return a list including

beta
A named vector of coefficients

iterations
A positive integer, iterative times in total.

stop.code
The stop code returned by nlm.

See Also

print.lm

Examples



  data(nhanes)
  fimllm <- fimlreg(age~., data=nhanes, iterlim= 40)
  a <- print(fimllm)
  a

[Package ILSE version 0.1.6 Index]