print.ilse {ILSE}	R Documentation

Print the Information of Iterative Least Square Estimation Regression When Covariates Include Missing Values

Description

print method for class "ilse"..

Usage


  ## S3 method for class 'ilse'
  print(object)

Arguments

object
an object of class "ilse".

Value

print the basic information of ilse estimation and algorithm and return a list including

beta
a named vector of coefficients

Bmat
a named matrix that summary the estimated beta in every iteration.

residuals
the residuals, that is response minus fitted values.

fitted.values
the fitted mean values.

d.fn
a nonnegative value, vlaue of relative variation rate of objective function value

d.par
a nonnegative value, relative variation rate of estimated parametric vector when algorithm stopped.

K
a positive integer, iterative times in total.

See Also

print.lm

Examples



data(nhanes)
NAlm1 <- ilse(age~., data=nhanes)
a <- print(NAlm1)
a

[Package ILSE version 0.1.6 Index]