Summarizing the inference information for ILSE or FIML methods

summary method for class "ilse" or "fiml".

summary(object, Nbt=20)

  # S3 method for ilse
summary(object, Nbt=20)

  # S3 method for fiml
summary(object, Nbt=20)

  ##
  Fitted.values(object)
  ##
  Residuals(object)

Arguments

object
an object of class "ilse".

Nbt
an positive integer, the repeated times of bootstrap to eatimate covariance matrix of regression coefficient.

Value

The function summary.ilse computes and returns a named matrix of summary statistics of the fitted linear model given in object by ILSE or FIML methods. The function Fitted.values return a vector, fitted repsonse vlaues. The function Residuals return a vector, residuals.

See also

Examples

# example one
data(nhanes)
NAlm <- ilse(age~., data=nhanes)
summary(NAlm, Nbt=5)

#> ===================Start bootstrapping================

#> ===================Finish bootstrapping================

#>                Estimate  std. Error   Z value     Pr(>|Z|)
#> (Intercept)  2.06942664 0.420946545  4.916127 8.827318e-07
#> bmi         -0.11656313 0.040330142 -2.890224 3.849679e-03
#> hyp          0.62109971 0.456887171  1.359416 1.740148e-01
#> chl          0.01056364 0.004463097  2.366887 1.793843e-02

fimllm <- fimlreg(age~., data=nhanes)
summary(fimllm, Nbt = 5)

#> ===================Start bootstrapping================

#> ===================Finish bootstrapping================

#>              Estimate std. Error    Z value     Pr(>|Z|)
#> intercept  2.37216291 0.59046199  4.0174693 5.882649e-05
#> bmi       -0.13518966 0.44922497 -0.3009398 7.634604e-01
#> hyp        0.46886921 1.04053468  0.4506041 6.522749e-01
#> chl        0.01253721 0.06695098  0.1872595 8.514572e-01