summary.fiml {ILSE}	R Documentation

Summarizing full-information Maximum Likelihood Estimation Regression Whether Covariates Include Missing Values or Response Include Missing Values.

Description

summary method for class "fiml"

Usage


  ## S3 method for class 'fiml'
  summary(object, Nbt=100)

Arguments

object
an object of class "fiml".

Nbt
an positive integer, the repeated times of bootstrap to eatimate covariance matrix of regression coefficient.

Value

The function summary.fiml computes and returns a named matrix of summary statistics of the fitted linear model given in object by full-information maximum likelihood.

See Also

summary.lm

Examples


  # example one
  data(nhanes)
  fimllm <- fimlreg(age~., data=nhanes, iterlim= 40)
  summary(fimllm, Nbt = 50)

[Package ILSE version 0.1.6 Index]