summary.ilse {ILSE}	R Documentation

Summarizing Iterative Least Square Estimation Regression When Covariates Include Missing Values

Description

summary method for class "ilse"

Usage


  ## S3 method for class 'ilse'
  summary(object, Nbt=100)
  ##
  Fitted.values(object)
  ##
  Residuals(object)


Arguments

object
an object of class "ilse".

Nbt
an positive integer, the repeated times of bootstrap to eatimate covariance matrix of regression coefficient.

Value

The function summary.ilse computes and returns a named matrix of summary statistics of the fitted linear model given in object by corrected least square estimation. The function Fitted.values return a vector, fitted repsonse vlaues. The function Residuals return a vector, residuals.

See Also

summary.lm fitted.vlaues residuals

Examples


# example one
data(nhanes)
NAlm <- ilse(age~., data=nhanes)
summary(NAlm, Nbt=50)

[Package ILSE version 0.1.6 Index]